Create `analysis`, `data`, `model`, `preprocessing`, and `view` packages. Setup for API reference generation.
