+++ src/genophenocorr/__init__.py
@@ -1,12 +1,11 @@
 # __init__.py
 
-from .patient import Patient, PatientCreator, PhenopacketPatientCreator, BasicPatientCreator
-from .phenotype import Phenotype, PhenotypeCreator, PhenotypeValidationException
-from .cohort import Cohort, CohortAnalysis
+from . import analysis
 from . import data
-from .protein import ProtCachingFunctionalAnnotator, ProteinAnnotationCache, ProteinFeature, ProteinMetadata, ProteinMetadataService, SimpleProteinFeature, UniprotProteinMetadataService
-from .variant import VarCachingFunctionalAnnotator, Variant, VariantAnnotationCache, VariantCoordinateFinder, VariantCoordinates, VepFunctionalAnnotator, PhenopacketVariantCoordinateFinder, FunctionalAnnotator, TranscriptAnnotation, BasicVariantCoordinateFinder
+from . import model
+from . import preprocessing
+from . import view
+
 from .constants import VariantEffect
-from .predicate import ProtFeaturePredicate, ProtFeatureTypePredicate, ExonPredicate, VariantPredicate, HPOPresentPredicate, VariantEffectPredicate
 
 __version__ = "0.1.0"
