+++ src/genophenocorr/analysis/__init__.py
@@ -1,0 +1,7 @@
+from . import predicate
+
+from ._analyzers import CohortAnalysis
+
+__all__ = [
+    'CohortAnalysis'
+]
