+++ src/genophenocorr/analysis/predicate/__init__.py
@@ -1,0 +1,9 @@
+from ._api import SimplePredicate, PatientCategory, PolyPredicate
+from ._all_predicates import HPOPresentPredicate, VariantEffectPredicate, VariantPredicate, ExonPredicate, ProtFeaturePredicate, ProtFeatureTypePredicate
+from ._all_predicates import HOMOZYGOUS, HETEROZYGOUS, NO_VARIANT
+
+__all__ = [
+    'SimplePredicate', 'PatientCategory', 'PolyPredicate',
+    'HPOPresentPredicate', 'VariantEffectPredicate', 'VariantPredicate', 'ExonPredicate', 'ProtFeaturePredicate', 'ProtFeatureTypePredicate',
+    'HOMOZYGOUS', 'HETEROZYGOUS', 'NO_VARIANT'
+]
