+++ src/genophenocorr/data/__init__.py
@@ -1,1 +1,3 @@
 from ._toy import get_toy_cohort
+
+__all__ = ['get_toy_cohort']
