+++ src/genophenocorr/model/__init__.py
@@ -1,0 +1,11 @@
+from ._cohort import Patient, Cohort
+from ._phenotype import Phenotype
+from ._protein import FeatureInfo, FeatureType, ProteinFeature, ProteinMetadata
+from ._variant import VariantCoordinates, TranscriptAnnotation, Variant
+
+__all__ = [
+    'Phenotype',
+    'FeatureInfo', 'FeatureType', 'ProteinFeature', 'ProteinMetadata',
+    'VariantCoordinates', 'TranscriptAnnotation', 'Variant',
+    'Patient', 'Cohort'
+]
