--- src/genophenocorr/patient/__init__.py
@@ -1,2 +1,0 @@
-from ._patient_data import Patient
-from ._annotators import PatientCreator, PhenopacketPatientCreator, BasicPatientCreator
