--- src/genophenocorr/phenotype/__init__.py
@@ -1,2 +1,0 @@
-from ._annotators import PhenotypeCreator, PhenotypeValidationException
-from ._phenotype_data import Phenotype
