--- src/genophenocorr/predicate/__init__.py
@@ -1,3 +1,0 @@
-from ._api import SimplePredicate, PolyPredicate, PatientCategory
-from ._all_predicates import VariantEffectPredicate, HPOPresentPredicate, VariantPredicate, ExonPredicate, ProtFeatureTypePredicate, ProtFeaturePredicate, PatientCategory, \
-    HOMOZYGOUS, HETEROZYGOUS, NO_VARIANT
