+++ src/genophenocorr/preprocessing/test_variant.py
@@ -1,12 +1,18 @@
-
 import pytest
+
+from pkg_resources import resource_filename
+
 from google.protobuf.json_format import Parse
+
 # pyright: reportGeneralTypeIssues=false
 from phenopackets import Phenopacket, GenomicInterpretation
 
-from ._annotators import VariantCoordinates, verify_start_end_coordinates, PhenopacketVariantCoordinateFinder, VepFunctionalAnnotator, VariantAnnotationCache, VarCachingFunctionalAnnotator
-from genophenocorr.protein import UniprotProteinMetadataService, ProteinAnnotationCache, ProtCachingFunctionalAnnotator
-from pkg_resources import resource_filename
+from genophenocorr.model import VariantCoordinates
+
+from ._phenopacket import PhenopacketVariantCoordinateFinder
+from ._protein import ProteinAnnotationCache, ProtCachingFunctionalAnnotator
+from ._uniprot import UniprotProteinMetadataService
+from ._variant import verify_start_end_coordinates, VepFunctionalAnnotator, VariantAnnotationCache, VarCachingFunctionalAnnotator
 
 
 @pytest.mark.parametrize('contig, start, end, ref, alt, chlen, expected',
@@ -105,6 +111,8 @@ def oldfile_cache_annotator(variant_annotator):
     vac = VariantAnnotationCache(data)
     return VarCachingFunctionalAnnotator(vac, variant_annotator)
 
+
+@pytest.mark.skip("Skipped until new Pickle files are generated")
 def test_cache_from_older_file(oldfile_cache_annotator, pp_vc_finder, variant_annotator):
     fname = resource_filename(__name__, 'test_data/missense_test.json')
     gi = read_genomic_interpretation_json(fname)
