--- src/genophenocorr/protein/__init__.py
@@ -1,2 +1,0 @@
-from ._annotators import ProteinMetadataService, UniprotProteinMetadataService, ProtCachingFunctionalAnnotator, ProteinAnnotationCache
-from ._protein_data import ProteinMetadata, ProteinFeature, SimpleProteinFeature, FeatureType, FeatureInfo
