--- src/genophenocorr/variant/__init__.py
@@ -1,3 +1,0 @@
-from ._variant_data import VariantCoordinates, Variant, TranscriptAnnotation
-from ._models import FunctionalAnnotator, VariantCoordinateFinder
-from ._annotators import VariantAnnotationCache, VepFunctionalAnnotator, VarCachingFunctionalAnnotator, PhenopacketVariantCoordinateFinder, BasicVariantCoordinateFinder
