+++ src/genophenocorr/view/__init__.py
@@ -1,0 +1,5 @@
+from ._cohort import CohortViewer
+
+__all__ = [
+    'CohortViewer'
+]
